abap-code-studio/cloud-backend/requirements.txt
# ABAP Code Studio — Cloud Backend Dependencies

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# Database
asyncpg>=0.29.0

# Auth
pyjwt[crypto]>=2.8.0

# Data validation
pydantic>=2.6.0

# HTTP client (for GitHub/GitLab API)
httpx>=0.27.0

# AI SDKs
anthropic>=0.25.0

# Utilities
python-multipart>=0.0.9
pyyaml>=6.0
redis>=5.0
